+++ services/blog-content/src/main/java/de/tomsblog/blogcontent/application/service/PostNotFoundException.java
@@ -1,6 +1,7 @@
 package de.tomsblog.blogcontent.application.service;
 
 import de.tomsblog.blogcontent.domain.model.PostId;
+import de.tomsblog.blogcontent.domain.model.Slug;
 
 public class PostNotFoundException extends RuntimeException {
 
@@ -9,6 +10,11 @@ public class PostNotFoundException extends RuntimeException {
     public PostNotFoundException(PostId postId) {
         super("Post not found: " + postId.asString());
         this.postId = postId;
+    }
+
+    public PostNotFoundException(Slug slug) {
+        super("Post not found: " + slug.value());
+        this.postId = null;
     }
 
     public PostId getPostId() {